+++ ios/adyen_checkout/Sources/adyen_checkout/utils/ConfigurationMapper.swift
@@ -43,11 +43,16 @@ extension DropInConfigurationDTO {
         if let cashAppPayConfigurationDTO {
             dropInConfiguration.cashAppPay = DropInComponent.CashAppPay(redirectURL: URL(string: cashAppPayConfigurationDTO.returnUrl)!)
         }
-        
+
         if let twintConfigurationDTO {
             dropInConfiguration.actionComponent.twint = .init(callbackAppScheme: twintConfigurationDTO.iosCallbackAppScheme)
         }
-        
+
+
+        if let threeDS2ConfigurationDTO {
+            dropInConfiguration.actionComponent.threeDS = buildThreeDS2Configuration(from: threeDS2ConfigurationDTO)
+        }
+
         dropInConfiguration.style = AdyenAppearance.dropInStyle
 
         return dropInConfiguration
@@ -100,7 +105,6 @@ extension DropInConfigurationDTO {
 
         return billingAddressConfiguration
     }
-
 }
 
 extension FieldVisibility {
@@ -127,7 +131,7 @@ extension DropInConfigurationDTO {
 }
 
 extension CardConfigurationDTO {
-    func mapToCardComponentConfiguration(shopperLocale: String?) -> CardComponent.Configuration {
+    func mapToCardComponentConfiguration(shopperLocale: String?, threeDS2ConfigurationDTO: ThreeDS2ConfigurationDTO? = nil) -> CardComponent.Configuration {
         let cardComponentStyle = AdyenAppearance.cardComponentStyle
         let localizationParameters = shopperLocale != nil ? LocalizationParameters(enforcedLocale: shopperLocale!) : nil
         let koreanAuthenticationMode = kcpFieldVisibility.toCardFieldVisibility()
@@ -177,6 +181,15 @@ extension CardConfigurationDTO {
         }
 
         return billingAddressConfiguration
+    }
+
+    private func buildCardThreeDS2Configuration(from threeDS2ConfigurationDTO: ThreeDS2ConfigurationDTO) -> AdyenActionComponent.Configuration.ThreeDS {
+        guard let url = URL(string: threeDS2ConfigurationDTO.requestorAppURL) else {
+            fatalError("Invalid URL string for requestorAppURL: \(threeDS2ConfigurationDTO.requestorAppURL)")
+        }
+        return AdyenActionComponent.Configuration.ThreeDS(
+            requestorAppURL: url
+        )
     }
 }
 
@@ -237,7 +250,7 @@ extension InstantPaymentConfigurationDTO {
     }
 }
 
-private func buildAdyenContext(environment: Environment, clientKey: String, amount: AmountDTO?, analyticsOptionsDTO: AnalyticsOptionsDTO, countryCode: String?) throws -> AdyenContext {
+func buildAdyenContext(environment: Environment, clientKey: String, amount: AmountDTO?, analyticsOptionsDTO: AnalyticsOptionsDTO, countryCode: String?) throws -> AdyenContext {
     let environment = environment.mapToEnvironment()
     let apiContext = try APIContext(
         environment: environment,
@@ -326,4 +339,9 @@ extension ActionComponentConfigurationDTO {
             countryCode: nil
         )
     }
+}
+
+func buildThreeDS2Configuration(from threeDS2ConfigurationDTO: ThreeDS2ConfigurationDTO) -> AdyenActionComponent.Configuration.ThreeDS {
+    guard let url = URL(string: threeDS2ConfigurationDTO.requestorAppURL) else { return .init() }
+    return .init(requestorAppURL: url)
 }